refactor: refactored InlineImageNode - spans for correct HTML
