+++ packages/lexical-playground/src/nodes/InlineImageNode/InlineImageNode.css
