+++ packages/lexical-playground/src/nodes/InlineImageNode/InlineImageNode.tsx
